courseMaterial: Implement deleting of materials

+++ courseMaterial/courseMaterial.go
@@ -5,6 +5,7 @@ import (
 	"database/sql"
 	"fmt"
 	"io"
+	"time"
 
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/volatiletech/sqlboiler/v4/boil"
@@ -13,7 +14,7 @@ import (
 	"learningbay24.de/backend/models"
 )
 
-// GetMaterial takes an ID and returns a struct of the file with the corresponding ID
+// GetMaterialFromCourse takes an ID and returns a struct of the file with the corresponding ID
 func GetMaterialFromCourse(db *sql.DB, fileId int) (*models.File, error) {
 	cm, err := models.FindFile(context.Background(), db, fileId)
 	if err != nil {
@@ -22,7 +23,7 @@ func GetMaterialFromCourse(db *sql.DB, fileId int) (*models.File, error) {
 	return cm, err
 }
 
-// GetMaterials takes a courseID and returns a slice of files associated with it
+// GetMaterialSliceFromCourse takes a courseID and returns a slice of files associated with it
 func GetMaterialSliceFromCourse(db *sql.DB, courseId int) (models.FileSlice, error) {
 	files, err := models.Files(
 		qm.From(models.TableNames.CourseHasFiles),
@@ -67,9 +68,8 @@ func CreateMaterial(dbHandle *sql.DB, fileName string, uri string, uploaderId, c
 	return nil
 }
 
-// DeactivateMaterial takes both course-ID and file-ID and deactivates the chosen material
-// Sets deactivation-timer and updates database
-func DeactivateMaterialFromCourse(db *sql.DB, courseId, fileId int) error {
+// DeleteMaterialFromCourse takes both course-ID and file-ID and deactivates the chosen material
+func DeleteMaterialFromCourse(db *sql.DB, courseId, fileId int) error {
 	tx, err := db.BeginTx(context.Background(), nil)
 	if err != nil {
 		return err
@@ -82,7 +82,39 @@ func DeactivateMaterialFromCourse(db *sql.DB, courseId, fileId int) error {
 		}
 	}
 
-	_, err = cm.Delete(context.Background(), tx, false)
+	// If more than 10 minutes have passed: soft-delete instead of hard-delete
+	curTime := time.Now()
+	diff := curTime.Sub(cm.CreatedAt)
+	if diff.Minutes() > 10 {
+		_, err = cm.Delete(context.Background(), tx, false)
+		if err != nil {
+			if e := tx.Rollback(); e != nil {
+				return fmt.Errorf("fatal: unable to rollback transaction on error: %s; %s", err.Error(), e.Error())
+			}
+		}
+
+		chf, err := models.FindCourseHasFile(context.Background(), tx, courseId, fileId)
+		if err != nil {
+			if e := tx.Rollback(); e != nil {
+				return fmt.Errorf("fatal: unable to rollback transaction on error: %s; %s", err.Error(), e.Error())
+			}
+		}
+
+		_, err = chf.Delete(context.Background(), tx, false)
+		if err != nil {
+			if e := tx.Rollback(); e != nil {
+				return fmt.Errorf("fatal: unable to rollback transaction on error: %s; %s", err.Error(), e.Error())
+			}
+		}
+
+		if e := tx.Commit(); e != nil {
+			return fmt.Errorf("fatal: unable to commit transaction on error: %s; %s", err, e)
+		}
+
+		return nil
+	}
+
+	_, err = cm.Delete(context.Background(), tx, true)
 	if err != nil {
 		if e := tx.Rollback(); e != nil {
 			return fmt.Errorf("fatal: unable to rollback transaction on error: %s; %s", err.Error(), e.Error())
@@ -96,7 +128,7 @@ func DeactivateMaterialFromCourse(db *sql.DB, courseId, fileId int) error {
 		}
 	}
 
-	_, err = chf.Delete(context.Background(), tx, false)
+	_, err = chf.Delete(context.Background(), tx, true)
 	if err != nil {
 		if e := tx.Rollback(); e != nil {
 			return fmt.Errorf("fatal: unable to rollback transaction on error: %s; %s", err.Error(), e.Error())
@@ -110,8 +142,8 @@ func DeactivateMaterialFromCourse(db *sql.DB, courseId, fileId int) error {
 	return nil
 }
 
-// DeactivateAllMaterials takes the ID of a course and deactivates all files associated with it
-func DeactivateAllMaterialsFromCourse(db *sql.DB, courseId int) error {
+// DeleteAllMaterialsFromCourse takes the ID of a course and deactivates all files associated with it
+func DeleteAllMaterialsFromCourse(db *sql.DB, courseId int, hardDelete bool) error {
 	tx, err := db.BeginTx(context.Background(), nil)
 	if err != nil {
 		return err
@@ -125,7 +157,15 @@ func DeactivateAllMaterialsFromCourse(db *sql.DB, courseId int) error {
 
 		return err
 	}
-	materials.DeleteAll(context.Background(), tx, false)
+
+	_, err = materials.DeleteAll(context.Background(), tx, hardDelete)
+	if err != nil {
+		if e := tx.Rollback(); e != nil {
+			return fmt.Errorf("fatal: unable to rollback transaction on error: %s; %s", err, e)
+		}
+
+		return err
+	}
 
 	if e := tx.Commit(); e != nil {
 		return fmt.Errorf("fatal: unable to commit transaction on error: %s; %s", err, e)
@@ -133,14 +173,18 @@ func DeactivateAllMaterialsFromCourse(db *sql.DB, courseId int) error {
 	return nil
 }
 
-// UpdateMaterial takes the ID of an existing file and name and overwrites the corresponding string with the new one
-func RenameMaterialFromCourse(db *sql.DB, fileId int, name string) error {
+// RenameMaterialFromCourse takes the ID of an existing file and name and overwrites the corresponding string with the new one
+func RenameMaterialFromCourse(db *sql.DB, fileId int, fileName string) error {
+	if fileName == "" {
+		return fmt.Errorf("Invalid value for variable fileName: String can't be empty!")
+	}
+
 	cm, err := models.FindFile(context.Background(), db, fileId)
 	if err != nil {
 		return err
 	}
 
-	cm.Name = name
+	cm.Name = fileName
 
 	_, err = cm.Update(context.Background(), db, boil.Infer())
 	if err != nil {
